Optimize for Render: float32 precision, build script, dynamic port

+++ src/ml/similarity.py
@@ -5,5 +5,6 @@ class SimilarityEngine:
         pass
 
     def compute_similarity(self, vectors):
-        
+        # Convert to float32 to save memory (Render free tier has 512MB limit)
+        vectors = vectors.astype('float32')
         return cosine_similarity(vectors)
